+++ examples/capataz-simple-example/src/example2/Main.hs
@@ -25,7 +25,8 @@ main :: IO ()
 main = do
   n       <- getRecord "Counter spawner"
 
-  capataz <- forkCapataz -- (1)
+  capataz <- forkCapataz
+    "unix-process-capataz" -- (1)
     (set supervisorRestartStrategyL OneForOne -- (2)
                                               . set onSystemEventL pPrint)                -- (3)