+++ src/Control/Concurrent/Capataz/Internal/Util.hs
@@ -163,8 +163,7 @@ sendSyncControlMsg SupervisorEnv { supervisorNotify } mkCtrlMsg = do
 -- | Utility function to transform a "CapatazOptions" record to a
 -- "SupervisorOptions" record.
 capatazOptionsToSupervisorOptions :: CapatazOptions -> SupervisorOptions
-capatazOptionsToSupervisorOptions CapatazOptions {..} =
-  SupervisorOptions {supervisorName = "capataz-root-supervisor", ..}
+capatazOptionsToSupervisorOptions CapatazOptions {..} = SupervisorOptions {..}
 
 -- | Utility function to transform a "SupervisorEnv" record to a
 -- "ParentSupervisorEnv" record; used on functions where supervision of